Prevent CensorRules being saved when Regexp is bad
We allow the regexp to be created from a user inputted string.

In some cases, we see rules that don’t have correct escaping, e.g:

    Regexp.new('foo]')
    # (irb):19: warning: regular expression has ']' without escape: /foo]/
    # => /foo]/

This warning behaviour is not configurable, so instead I’ve added
Warning.with_raised_warnings. This simply makes any call to Warning.warn
raise the given message instead of outputting it to stderr.

We then use the raised exception to massage the message into a
RegexpError, which we already have handling that presents this as an
ActiveModel error to the user.

Fixes https://github.com/mysociety/alaveteli/issues/5967